+++ internal/connectors/nats/nats.go
@@ -1,7 +1,9 @@
 package nats
 
 import (
+	"fmt"
 	"sync"
+	"time"
 
 	"github.com/nats-io/nats.go"
 
@@ -17,7 +19,17 @@ func New(cfg *env.AppConfig) (*nats.Conn, error) {
 	var err error
 
 	onceDefaultClient.Do(func() {
-		natsClient, err = nats.Connect(cfg.NatsDefaultURL)
+		natsClient, err = nats.Connect(cfg.NatsDefaultURL,
+			nats.ReconnectWait(2*time.Second),
+			nats.DisconnectErrHandler(func(nc *nats.Conn, err error) {
+				fmt.Printf("Nats client got disconnected!\n")
+			}),
+			nats.ReconnectHandler(func(nc *nats.Conn) {
+				fmt.Printf("Nats client got got reconnected to %v!\n", nc.ConnectedUrl())
+			}),
+			nats.ClosedHandler(func(nc *nats.Conn) {
+				fmt.Printf("Nats connection closed\n")
+			}))
 	})
 
 	return natsClient, err