+++ Changelog.md
@@ -1,3 +1,7 @@
+## 17.06.2024
+1. Added reconnect feature for nats client
+2. Added swagger 200, 400 responses for /alert handler
+
 ## 07.06.2024
 1. Optimized nats consumer worker
 2. Optimized alert handlers